Add support for django-s3-storage-folder
Fixes a bug when using django-s3-storage-folder, where filebrowser-
safe will not set the appropriate StorageMixin because it doesn't
recognize 's3_folder_storage.s3.DefaultStorage' as a valid value for
DEFAULT_FILE_STORAGE.

Specifically, it looks for
filebrowser_safe.storage.DefaultStorageMixin, which doesn't exist.

Instead of defining a storage.DefaultStorageMixin class (which would
have a misleading name), we just treat this as a special case and
use the storage.S3BotoStorageMixin.